# repository: Re-tune indexmap allocation strategy